+++ src/main/java/br/com/vitavault/exceptions/EstoqueException.java
@@ -1,0 +1,7 @@
+package br.com.vitavault.exceptions;
+
+public class EstoqueException extends Exception{
+    public EstoqueException(String message) {
+        super(message);
+    }
+}