+++ src/main/java/br/com/vitavault/exceptions/MovimentacaoEstoqueException.java
@@ -1,0 +1,7 @@
+package br.com.vitavault.exceptions;
+
+public class MovimentacaoEstoqueException extends Exception{
+    public MovimentacaoEstoqueException(String message) {
+        super(message);
+    }
+}